Clean up to prepare for open sourcing
Adding README. Removing To-do list and journal. Reorganizing directory
hierarchy. Deleting .pyc's. Deleting backup folder.
