--- TODO.md
@@ -1,63 +1,0 @@
-# TODO.md
-
-### Misc.
-
-* Scan design documents
-* Set up git, github **done, done**
-
-### Features
-
-* Add 'custom word' feature for Web
-* Add 'custom word' feature for Discord via DM
-
-### System Design
-
-* Start requirements
-* Design back end code / data structures / classes **progress**
-* Design interactions for different clients
-* Authentication research
-
-### Front End
-
-* PHP script to display results
-* PHP script to interpret JSON for Web
-* PHP script for game interface
-* HTML/CSS for Web interface
-* Flask methods to interpret JSON for Discord
-* Flask methods for game input **progress**
-
-### API
-
-* Flask setup **progress**
-* Contractualize JSON format for v.0.1
-* Versioning
-* GET new board **done**
-* POST new guess **done**
-
-### Backend
-
-* Separate functions  **progress**
-* Remove unneeded functions **progress**
-* Refactor legacy code **progress**
-* Write ping function **progress**
-* Write testing functions
-* Tell user when they completed the game (win or lose)
-* Track turns
-* Stats
-* Security / Auth / Privatizing game
-
-### Testing
-
-* Ping function **works**
-* Prove a word was picked
-* Prove a word was remembered
-* Prove a guess was validated
-* Prove a guess was played
-* Prove the keyboard is generated
-* Prove the keyboard is updated
-* Prove the keyboard is displayed
-* Prove the progress is generated
-* Prove the progress is updated
-* Prove the progress is displayed
-* Prove the game over screen is displayed
-* Prove you can't play someone elses' game
