+++ src/api/backend_setup.py
@@ -127,92 +127,3 @@ def start_up_game_backend(cache_id):
 
     return my_cache, common_words, all_words
 
-
-#     def test2():
-#       """main game logic. initializes several variables, then jumps into the main loop:
-#       1. compare a guess
-#       2. print the new board
-#       3. repeat."""
-#       guess_history = []
-#       index_map_history = []
-#       common_words, all_words = load_dicts_from_json()
-#       game_day = 0
-#       shared_id, shared_word = ask_user_new_word_or_word_from_id(common_words)
-#       if shared_word and shared_id:
-#           todays_word_game_id = shared_id
-#           todays_word = shared_word
-#       else:
-#           todays_word_game_id, todays_word = get_todays_word(common_words)
-#       if CHEATING: print("[cheating] todays word is", todays_word)
-#       key_map = create_keyboard_map()
-#       emoji_hash = create_emoji_hash()
-#       guesses = 0
-#   
-#       print("Welcome to Polywordle [version 0.2]")
-#   
-#       # Print the "blank" board with just empty squares.
-#       for x in range(guesses, 6):
-#           pretty_print_blank_lines(emoji_hash, 'white')
-#       pretty_print_keyboard(key_map)
-#   
-#       print("this is key_map")
-#       json_status_dict = {}
-#       json_status_dict['keyboard'] = key_map
-#       json_status_dict['game_id'] = 0
-#       json_status_dict['turn'] = guesses
-#       print(key_map)
-#       print("this is the json response")
-#       print(json_status_dict)
-#       print(json.dumps(json_status_dict, indent=4))
-#   
-#       while guesses < 6:
-#           invalid_guess = True
-#           while invalid_guess:
-#               current_guess = input(("Guess #"+ str(guesses+1)+ ": "))
-#               if len(current_guess) == 5:
-#                   pattern = re.compile("[A-Za-z]+")
-#                   if pattern.fullmatch(current_guess):
-#                       if current_guess.lower() in list(all_words.values()): break
-#                       else: print("Not in dictionary.")
-#                   else: print("Must only be letters.")
-#               else: print("Must be 5 letters.")
-#   
-#           current_guess = current_guess.lower()
-#   
-#           guesses += 1
-#           # note sure if this is right
-#           index_color_map, key_map = update_all(current_guess, todays_word, key_map, index_map_history)
-#           guess_history.append(current_guess)
-#   
-#           print("now printing the json response")
-#           json_status_dict['keyboard'] = key_map
-#           json_status_dict['index_color_map_history'] = index_map_history
-#           json_status_dict['turn'] = guesses
-#           print(json.dumps(json_status_dict, indent=4))
-#   
-#   
-#           # clear screen
-#           for x in range(50):print("")
-#           for x in range(guesses):
-#               pretty_print_index_color(index_map_history[x], guess_history[x], emoji_hash)
-#   
-#           for x in range(guesses, 6):
-#               pretty_print_blank_lines(emoji_hash, 'white')
-#   
-#           pretty_print_keyboard(key_map)
-#           if current_guess == todays_word:
-#               break
-#       if current_guess == todays_word:
-#           print(text_hash['green'], end='')
-#           print(performance_hash[guesses])
-#           print(text_hash['end'], end='')
-#       else:
-#           print(text_hash['red'], "Bummer! The word was ", todays_word.upper(), text_hash['end'], sep='')
-#       print(f"\nIf you feel this word is unfair, please open an issue at ", text_hash['yellow'], "https://GitHub.com/JoeBussard/PolyWordle", text_hash['end'], sep='')
-#   
-#       print(f"\nShare your score:\n")
-#       print(generate_share_text(guesses, index_map_history, emoji_hash, todays_word_game_id))
-#   
-#       exit(0)
-#   
-#   #test2()
